requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
streamlit
matplotlib
scikit-learn